refactor: rename application service files to remove _service suffix
- Rename complete_task_service.py to complete_task.py
- Rename create_task_service.py to create_task.py
- Rename get_task_service.py to get_task.py
- Rename list_tasks_service.py to list_tasks.py
- Update __init__.py imports to reflect new file names
- Update documentation in TUTORIAL.md
