+++ src/application/services/__init__.py
@@ -1,7 +1,7 @@
-from .create_task_service import CreateTaskService
-from .get_task_service import GetTaskService
-from .list_tasks_service import ListTasksService
-from .complete_task_service import CompleteTaskService
+from .create_task import CreateTaskService
+from .get_task import GetTaskService
+from .list_tasks import ListTasksService
+from .complete_task import CompleteTaskService
 
 __all__ = [
     "CreateTaskService",
